solutions/requirements.txt
# LangChain and LangGraph core dependencies
langchain>=0.1.0
langchain-openai>=0.0.5
langgraph>=0.0.20
langchain-core>=0.1.0
langchain-community>=0.0.10
langchain-anthropic

# Language Model providers
openai>=1.0.0
anthropic>=0.18.0

# Core Python dependencies
pydantic>=2.0.0
pyyaml>=6.0
aiofiles>=23.0.0
python-dotenv>=1.0.0

# Tool dependencies
tiktoken>=0.5.0
typing-extensions>=4.8.0
asyncio>=3.4.3

# EDA tool dependencies (for parsing outputs)
pyparsing>=3.0.0
regex>=2023.0.0

# Visualization and debugging
matplotlib>=3.7.0
networkx>=3.0

# Testing and development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
isort>=5.12.0

# Optional: For human-in-the-loop features
gradio>=4.0.0
streamlit>=1.28.0 

# Logging
colorlog>=6.7.0